requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
python-multipart
pydantic
boto3
PyJWT
passlib[bcrypt]
pandas
openpyxl
python-dotenv
pytest
